# services/ai/app/api/v1/accessibility.py
"""
Accessibility endpoints for AI Service
"""

from fastapi import APIRouter, HTTPException, UploadFile, File, Form
from fastapi.responses import JSONResponse
from pydantic import BaseModel
from typing import List, Optional
import time
from loguru import logger

router = APIRouter()

class AccessibilityRequest(BaseModel):
    content: str
    type: str  # "image", "text", "document"
    features: List[str]  # ["alt_text", "contrast", "readability", "screen_reader"]

class AccessibilityResponse(BaseModel):
    alt_text: Optional[str] = None
    contrast_score: Optional[float] = None
    readability_score: Optional[float] = None
    screen_reader_compatible: Optional[bool] = None
    suggestions: List[str]

class SceneDescriptionRequest(BaseModel):
    image_url: str
    detail_level: Optional[str] = "medium"  # "low", "medium", "high"

class SceneDescriptionResponse(BaseModel):
    description: str
    objects: List[dict]
    confidence: float
    processing_time: float

class ObjectDetectionRequest(BaseModel):
    image_url: str
    confidence_threshold: Optional[float] = 0.7
    max_objects: Optional[int] = 20

class ObjectDetectionResponse(BaseModel):
    objects: List[dict]
    processing_time: float

@router.post("/analyze")
async def analyze_accessibility(request: AccessibilityRequest):
    """Analyze content for accessibility compliance"""
    start_time = time.time()
    
    try:
        # TODO: Implement actual accessibility analysis
        # This is a placeholder implementation
        
        result = {
            "alt_text": "Generated alt text for image" if request.type == "image" else None,
            "contrast_score": 0.85,
            "readability_score": 0.78,
            "screen_reader_compatible": True,
            "suggestions": ["Increase font size", "Add descriptive alt text"]
        }
        
        processing_time = time.time() - start_time
        logger.info(f"Accessibility analysis completed in {processing_time:.3f}s")
        
        return result
        
    except Exception as e:
        logger.error(f"Accessibility analysis failed: {e}")
        raise HTTPException(status_code=500, detail="Accessibility analysis failed")

@router.post("/enhance")
async def enhance_accessibility(request: AccessibilityRequest):
    """Enhance content for better accessibility"""
    start_time = time.time()
    
    try:
        # TODO: Implement actual accessibility enhancement
        # This is a placeholder implementation
        
        result = {
            "message": "Accessibility enhancement completed",
            "enhanced_content": request.content,
            "improvements": ["Added alt text", "Improved contrast", "Enhanced readability"]
        }
        
        processing_time = time.time() - start_time
        logger.info(f"Accessibility enhancement completed in {processing_time:.3f}s")
        
        return result
        
    except Exception as e:
        logger.error(f"Accessibility enhancement failed: {e}")
        raise HTTPException(status_code=500, detail="Accessibility enhancement failed")

@router.get("/guidelines")
async def get_accessibility_guidelines():
    """Get accessibility guidelines and best practices"""
    return {
        "guidelines": [
            "Provide alt text for images",
            "Ensure sufficient color contrast",
            "Use semantic HTML",
            "Add keyboard navigation",
            "Provide text alternatives for audio/video",
            "Use clear and simple language",
            "Ensure logical reading order",
            "Provide sufficient time for interactions"
        ]
    }

@router.post("/scene-description")
async def describe_scene(request: SceneDescriptionRequest):
    """Generate scene description for accessibility"""
    start_time = time.time()
    
    try:
        # TODO: Implement actual scene description using AI models
        # This is a placeholder implementation
        
        result = {
            "description": "A person standing in front of a building with trees in the background",
            "objects": [
                {"name": "person", "confidence": 0.95, "location": "center"},
                {"name": "building", "confidence": 0.87, "location": "background"},
                {"name": "trees", "confidence": 0.92, "location": "background"}
            ],
            "confidence": 0.91,
            "processing_time": time.time() - start_time
        }
        
        logger.info(f"Scene description completed in {result['processing_time']:.3f}s")
        return result
        
    except Exception as e:
        logger.error(f"Scene description failed: {e}")
        raise HTTPException(status_code=500, detail="Scene description failed")

@router.post("/object-detection")
async def detect_objects(request: ObjectDetectionRequest):
    """Detect objects in image for accessibility"""
    start_time = time.time()
    
    try:
        # TODO: Implement actual object detection using AI models
        # This is a placeholder implementation
        
        result = {
            "objects": [
                {"name": "chair", "confidence": 0.89, "bbox": [100, 100, 200, 300]},
                {"name": "table", "confidence": 0.94, "bbox": [300, 200, 500, 400]},
                {"name": "lamp", "confidence": 0.76, "bbox": [50, 50, 100, 150]}
            ],
            "processing_time": time.time() - start_time
        }
        
        logger.info(f"Object detection completed in {result['processing_time']:.3f}s")
        return result
        
    except Exception as e:
        logger.error(f"Object detection failed: {e}")
        raise HTTPException(status_code=500, detail="Object detection failed")