# services/api/src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Vision Platform API
  description: |
    Complete Vision Platform API for Cross-language Multimodal Translator and Multimodal Accessibility Companion
    
    ## Features
    - **Translation Services**: Text, speech, and image translation with 50+ languages
    - **Accessibility Features**: Scene description, object detection, navigation assistance
    - **Real-time Communication**: WebSocket support for live translation and accessibility
    - **AI Integration**: OpenAI, Hugging Face, and Google Cloud AI services
    - **User Management**: Authentication, profiles, preferences, and analytics
    
    ## Authentication
    All API endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Free tier: 100 requests/hour
    - Premium tier: 1000 requests/hour
    - Admin tier: Unlimited
    
    ## Support
    For API support, contact: api-support@vision-platform.com
  version: 1.0.0
  contact:
    name: Vision Platform API Support
    email: api-support@vision-platform.com
    url: https://vision-platform.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.vision-platform.com
    description: Production server
  - url: https://staging-api.vision-platform.com
    description: Staging server
...
paths:
...
  /api/accessibility/scene-description:
    post:
      tags:
        - Accessibility
      summary: Scene description
      description: Describe the content of an image for visually impaired users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Image file to analyze
                detail:
                  type: string
                  enum: [basic, detailed, comprehensive]
                  default: detailed
                includeObjects:
                  type: boolean
                  default: true
                includeText:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Scene description generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SceneDescriptionResponse'
  /api/accessibility/object-detection:
    post:
      tags:
        - Accessibility
      summary: Object detection
      description: Detect and identify objects in an image
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Image file to analyze
                confidence:
                  type: number
                  minimum: 0.1
                  maximum: 1.0
                  default: 0.7
      responses:
        '200':
          description: Objects detected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectDetectionResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SceneDescriptionResponse:
      type: object
      properties:
        description:
          type: string
          description: Scene description
        objects:
          type: array
          items:
            $ref: '#/components/schemas/DetectedObject'
        text:
          type: array
          items:
            type: string
          description: Text found in image
        confidence:
          type: number
          minimum: 0
          maximum: 1
    DetectedObject:
      type: object
      properties:
        name:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        boundingBox:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            width:
              type: number
            height:
              type: number
    ObjectDetectionResponse:
      type: object
      properties:
        objects:
          type: array
          items:
            $ref: '#/components/schemas/DetectedObject'
        totalObjects:
          type: integer
          minimum: 0
...